requirements.txt
# Bot Framework
python-telegram-bot==20.7
python-telegram-bot[webhooks]==20.7

# Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
sqlite3  # Built-in for development

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Image Processing
Pillow==10.2.0
aiohttp==3.9.1

# Scheduling
apscheduler==3.10.4

# AI Integration
openai==1.12.0

# Timezone handling
pytz==2024.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
